Refactor user stats calculations and ensure integer output in game-end email template
Updated the calculation of 'daysPlayed' in SendGameEndEmailsCommand to cast the result to an integer. Modified the game-end email template to ensure 'days_played' and 'total_actions' are displayed as integers, enhancing data consistency and presentation in user communications.

+++ resources/views/emails/game-end.blade.php
@@ -174,11 +174,11 @@
             @if(isset($stats))
             <div class="stats-grid">
                 <div class="stat-card">
-                    <p class="stat-value">{{ $stats['days_played'] ?? 0 }}</p>
+                    <p class="stat-value">{{ (int) ($stats['days_played'] ?? 0) }}</p>
                     <p class="stat-label">дней в игре</p>
                 </div>
                 <div class="stat-card">
-                    <p class="stat-value">{{ $stats['total_actions'] ?? 0 }}</p>
+                    <p class="stat-value">{{ (int) ($stats['total_actions'] ?? 0) }}</p>
                     <p class="stat-label">действий выполнено</p>
                 </div>
             </div>
